# main-app/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Changeloger API
  description: |
    RESTful API for Changeloger - automated changelog generation platform.
    All endpoints return JSON. Authentication is via JWT in httpOnly cookies
    (set automatically by the OAuth flow).
  version: 1.0.0
  contact:
    name: Changeloger
    url: https://changeloger.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: https://app.changeloger.com/api
    description: Production
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: changeloger_token
    embedToken:
      type: apiKey
      in: path
      name: embedToken
    webhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
...
paths:
...
  /github/installation:
    get:
      tags: [Repositories]
      summary: GitHub App installation callback
      description: Handles the redirect after a user installs the GitHub App. Stores the installation and syncs repositories.
      security:
        - cookieAuth: []
      parameters:
        - name: installation_id
          in: query
          required: true
          schema:
            type: integer
        - name: setup_action
          in: query
          required: true
          schema:
            type: string
            enum: [install]
      responses:
        '302':
          description: Redirect to repositories page